dont remove vagrant insecure key for desktop virtualization

+++ script/ubuntu.sh
@@ -6,6 +6,10 @@ passwd -l ubuntu
 gpasswd -a ubuntu docker
 gpasswd -a ubuntu sudo
 echo 'ubuntu ALL=(ALL) NOPASSWD: ALL' > /etc/sudoers.d/ubuntu
-rm -vf ~ubuntu/.ssh/authorized_keys
 date > ~ubuntu/.created
+
+if [[ ! $PACKER_BUILDER_TYPE =~ virtualbox|parallels|vmware ]]; then
+  rm -vf ~ubuntu/.ssh/authorized_keys
+fi
+
 find /home/ubuntu | xargs ls -ltrhd
